DROID-2121 rename
