+++ app/src/main/java/com/anytypeio/anytype/di/feature/EditorDI.kt
@@ -186,7 +186,7 @@ interface EditorSubComponent {
     fun linkToObjectOrWebBuilder(): LinkToObjectOrWebSubComponent.Builder
 
     fun tagStatusObjectComponent(): TagOrStatusValueObjectComponent.Builder
-    fun optionObjectComponent(): OptionObjectComponent.Builder
+    fun optionObjectComponent(): CreateOrEditOptionObjectComponent.Builder
 }
