+++ app/src/main/res/navigation/nav_relations.xml
@@ -11,7 +11,7 @@
 
     <dialog
         android:id="@+id/optionScreen"
-        android:name="com.anytypeio.anytype.ui.relations.value.OptionFragment"
+        android:name="com.anytypeio.anytype.ui.relations.value.CreateOrEditOptionFragment"
         android:label="Option create or edit screen" />
 
 </navigation>
